backend/requirements.txt
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0